+++ types/protyle/util/compatibility.d.ts
@@ -1,0 +1,16 @@
+export declare const openByMobile: (uri: string) => void;
+export declare const readText: () => string | Promise<string>;
+export declare const writeText: (text: string) => void;
+export declare const copyPlainText: (text: string) => Promise<void>;
+export declare const getEventName: () => "click" | "touchstart";
+export declare const isOnlyMeta: (event: KeyboardEvent | MouseEvent) => boolean;
+export declare const isNotCtrl: (event: KeyboardEvent | MouseEvent) => boolean;
+export declare const isHuawei: () => boolean;
+export declare const isIPhone: () => boolean;
+export declare const isIPad: () => boolean;
+export declare const isMac: () => boolean;
+export declare const isInAndroid: () => boolean;
+export declare const isInIOS: () => any;
+export declare const updateHotkeyTip: (hotkey: string) => string;
+export declare const getLocalStorage: (cb: () => void) => void;
+export declare const setStorageVal: (key: string, val: any, cb?: () => void) => void;